Add files via upload
This repository contains two machine learning projects, each focused on a different classification problem.

1. Titanic Classification
This project involves building a predictive model to classify passengers of the Titanic based on various features like age, gender, and class. The goal is to predict whether a passenger survived the tragic event. The notebook covers data exploration, preprocessing, feature engineering, model selection, and evaluation.

2. Email Classification
In this project, the focus is on classifying emails as either spam or non-spam (ham). The notebook walks through data preprocessing, feature extraction using techniques like TF-IDF, and the development of a machine learning model to accurately categorize the emails.